@@ -1,18 +1,22 @@
 {
-    "version": "1.0.0",
-    "name": "si-kolayout",
-    "author": "Poul K. Sorensen <pks@s-innovations.net> (http://twitter.com/pksorensen)",
-    "homepage": "https://www.s-innovations.net",
-    "bugs": {
-        "url": "https://github.com/s-innovations/si-appbuilder/issues",
-        "email": "bugs@s-innovations.com"
-    },
-    "devDependencies": {
-        "@types/knockout": "3.4.53",
-        "typescript": "2.8.3"
-    },
-    "files": [
-        "artifacts"
-    ],
-    "typings": "artifacts/typings/index.d.ts"
+  "version": "1.0.3",
+  "name": "si-kolayout",
+  "author": "Poul K. Sorensen <pks@s-innovations.net> (http://twitter.com/pksorensen)",
+  "homepage": "https://www.s-innovations.net",
+  "bugs": {
+    "url": "https://github.com/s-innovations/si-appbuilder/issues",
+    "email": "bugs@s-innovations.com"
+  },
+  "devDependencies": {
+"knockout": "3.5.0",
+    "typescript": "3.4.5"
+  },
+  "files": [
+    "artifacts"
+  ],
+  "scripts": {
+    "tsc": "node_modules/.bin/tsc",
+    "prepare": "npm run tsc"
+  },
+  "typings": "artifacts/typings/index.d.ts"
 }
